refactor (lottoService, lottos) 1. LottoService로 lottos, statistics 메소드 내부의 로컬 변수로 옮김 2. Lottos 일급객체로 분리
