+++ src/main/java/step3/LottoApplication.java
@@ -1,7 +1,10 @@
 package step3;
 
 import java.util.List;
+import java.util.Map;
+import step3.domain.Lotto;
 import step3.domain.LottoService;
+import step3.domain.Lottos;
 import step3.views.Input;
 import step3.views.Output;
 
@@ -10,6 +13,7 @@ public class LottoApplication {
 
         Input input = new Input();
         Output output = new Output();
+        Lottos lottos = new Lottos();
         LottoService lottoService = new LottoService();
 
         output.purchase();
@@ -17,15 +21,16 @@ public class LottoApplication {
         int purchasingNumber = lottoService.calculateLottoCount(money);
 
         output.generateLottos(purchasingNumber);
-        lottoService.generateLottos(purchasingNumber);
+        List<Lotto> lottoList = lottos.generateLottos(purchasingNumber);
 
         output.winnerNumbers();
         String winnerNumbersWithComma = input.inputString();
         List<Integer> winnerNumbers = lottoService.gainWinnerNumbers(winnerNumbersWithComma);
-        lottoService.matchWinningNumbers(winnerNumbers);
+        lottoService.matchWinningNumbers(lottoList, winnerNumbers);
 
-        output.statistic(lottoService.getStatistics(), lottoService.statisticLottos(money));
-
+        Map<Integer, Integer> statistics = lottoService.calculateWinningBallsEachLotto(lottoList);
+        double returnOnInvestmentRate = lottoService.statisticLottos(statistics, money);
+        output.statistic(statistics, returnOnInvestmentRate);
 
     }
 }
