+++ src/main/java/step3/domain/LottoService.java
@@ -1,6 +1,5 @@
 package step3.domain;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -11,36 +10,28 @@ import step3.enums.Award;
 public class LottoService {
 
     private final static int LOTTO_PRICE = 1000;
-    private final List<Lotto> lottos;
-    private final Map<Integer, Integer> statistics;
-
-    public LottoService() {
-        lottos = new ArrayList<>();
-        statistics = new HashMap<>();
-        initStatistics();
-    }
 
     public int calculateLottoCount(int money) {
         return money / LOTTO_PRICE;
     }
 
-    public List<Lotto> generateLottos(int count) {
-        for (int i = 0; i < count; i++) {
-            lottos.add(new Lotto());
-        }
-        return lottos;
-    }
-
-    public void matchWinningNumbers(List<Integer> winningNumbers) {
+    public void matchWinningNumbers(List<Lotto> lottos, List<Integer> winningNumbers) {
         lottos.forEach(lotto -> {
             lotto.match(winningNumbers);
         });
     }
 
-    public double statisticLottos(int money) {
+    public Map<Integer, Integer> calculateWinningBallsEachLotto(List<Lotto> lottos) {
+        Map<Integer, Integer> statistics = initStatistics();
+
         lottos.forEach(lotto -> {
             statistics.computeIfPresent(lotto.getMatchCount(), (k, v) -> v + 1);
         });
+        return statistics;
+    }
+
+    public double statisticLottos(Map<Integer, Integer> statistics, int money) {
+
         double sum = statistics.get(Award.THREE.getCount()) * Award.THREE.getAmount();
         sum += statistics.get(Award.FOUR.getCount()) * Award.FOUR.getAmount();
         sum += statistics.get(Award.FIVE.getCount()) * Award.FIVE.getAmount();
@@ -48,15 +39,15 @@ public class LottoService {
         return Math.floor(sum / money * 100) / 100;
     }
 
-    public Map<Integer, Integer> getStatistics() {
-        return statistics;
-    }
+    private Map<Integer, Integer> initStatistics() {
+        Map<Integer, Integer> statistics = new HashMap<>();
 
-    private void initStatistics() {
         statistics.put(Award.THREE.getCount(), 0);
         statistics.put(Award.FOUR.getCount(), 0);
         statistics.put(Award.FIVE.getCount(), 0);
         statistics.put(Award.SIX.getCount(), 0);
+
+        return statistics;
     }
 
     public List<Integer> gainWinnerNumbers(String numbersWithComma) {
