+++ src/main/java/step3/views/Output.java
@@ -18,12 +18,12 @@ public class Output {
         System.out.println(count + Message.COUNT.getMessage());
     }
 
-    public void statistic(Map<Integer, Integer> statistics, double statisticLottos) {
+    public void statistic(Map<Integer, Integer> statistics, double returnOnInvestmentRate) {
         System.out.println("\n" + Message.STATISTICS.getMessage());
         System.out.println(Message.MATCH_THREE.resultMatchNumber(statistics.get(Award.THREE.getCount())));
         System.out.println(Message.MATCH_FOUR.resultMatchNumber(statistics.get(Award.FOUR.getCount())));
         System.out.println(Message.MATCH_FIVE.resultMatchNumber(statistics.get(Award.FIVE.getCount())));
         System.out.println(Message.MATCH_SIX.resultMatchNumber(statistics.get(Award.SIX.getCount())));
-        System.out.println(Message.STATISTICS_RESULT.resultStatistic(statisticLottos));
+        System.out.println(Message.STATISTICS_RESULT.resultStatistic(returnOnInvestmentRate));
     }
 }
