+++ src/test/java/step3/LottoServiceTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
 import step3.domain.LottoService;
 
 class LottoServiceTest {
@@ -16,12 +15,6 @@ class LottoServiceTest {
     @CsvSource(value = {"14000:14", "1000:1", "3000:3"}, delimiter = ':')
     void calculateLottoCount(int input, int expected) {
         Assertions.assertEquals(expected, lottoService.calculateLottoCount(input));
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = {1, 2, 3})
-    void generateLottos(int input) {
-        Assertions.assertEquals(input, lottoService.generateLottos(input).size());
     }
 
 }
